ApiCrypt代码整理
